트랜젝션, db 연결 방식 수정

+++ src/modules/auth/auth.service.ts
@@ -1,17 +1,27 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { Inject, Injectable, Logger } from '@nestjs/common';
 import { type FastifyRequest } from 'fastify';
+import { type NodePgDatabase } from 'drizzle-orm/node-postgres';
 import { plainToInstance } from 'class-transformer';
 import bcrypt from 'bcrypt';
 
 import { UserRepository } from '@/modules/user/user.repository';
 import { AuthUtil } from './auth.util';
-import { NativeSignInRequestDto, RefreshRequestDto } from './dto/request';
+import {
+  CheckEmailRequestDto,
+  NativeSignInRequestDto,
+  NativeSignUpRequestDto,
+  RefreshRequestDto,
+  SignOutRequestDto,
+} from './dto/request';
 import { SignInResponseDto } from './dto/response';
 import { CustomException } from '@/common/api/exception/global.exception';
+import * as schema from '@/common/database/schemas';
 
 @Injectable()
 export class AuthService {
   constructor(
+    @Inject('DRIZZLE_CONNECTION')
+    private readonly db: NodePgDatabase<typeof schema>,
     private readonly userRepository: UserRepository,
     private readonly authUtil: AuthUtil,
   ) {}
@@ -22,6 +32,7 @@ export class AuthService {
     requestDto: NativeSignInRequestDto,
   ): Promise<SignInResponseDto> {
     const signInResult = await this.userRepository.findNativeSignInInfoByEmail(
+      this.db,
       requestDto.email,
     );
 
@@ -47,6 +58,7 @@ export class AuthService {
       );
 
     await this.userRepository.createRefreshToken(
+      this.db,
       signInResult.userId,
       tokenResponse.refreshToken,
       refreshTokenExpiredAt,
@@ -74,6 +86,7 @@ export class AuthService {
 
     const formalRefreshTokenInfo =
       await this.userRepository.findRefreshTokenIdByUserIdAndToken(
+        this.db,
         formalTokenInfo.userId,
         formalTokenInfo.refreshToken,
       );
@@ -83,6 +96,7 @@ export class AuthService {
     }
 
     const refreshResult = await this.userRepository.findRefreshInfoByUserId(
+      this.db,
       formalTokenInfo.userId,
     );
 
@@ -102,6 +116,7 @@ export class AuthService {
       );
 
     await this.userRepository.updateRefreshToken(
+      this.db,
       formalRefreshTokenInfo.refreshTokenId,
       tokenResponse.refreshToken,
       refreshTokenExpiredAt,
@@ -123,12 +138,84 @@ export class AuthService {
     });
   }
 
-  async signOut(request: FastifyRequest): Promise<void> {
-    const signOutInfo = await this.authUtil.getFormalRefreshInfo(request);
+  async signOut(requestDto: SignOutRequestDto): Promise<void> {
+    if (!requestDto.refreshToken) return;
+
     await this.userRepository.deleteRefreshTokenByToken(
-      signOutInfo.refreshToken,
+      this.db,
+      requestDto.refreshToken,
+    );
+  }
+
+  async checkEmail(requestDto: CheckEmailRequestDto): Promise<void> {
+    const isDuplicate = await this.userRepository.existsByEmail(
+      this.db,
+      requestDto.email,
     );
 
-    this.log.log(`Sign Out successfully for user ID: ${signOutInfo.userId}`);
+    if (isDuplicate) {
+      throw new CustomException('DUPLICATE_RESOURCE');
+    }
+  }
+
+  async nativeSignUp(
+    requestDto: NativeSignUpRequestDto,
+  ): Promise<SignInResponseDto> {
+    const isDuplicate = await this.userRepository.existsByEmail(
+      this.db,
+      requestDto.email,
+    );
+
+    if (isDuplicate) {
+      throw new CustomException('DUPLICATE_RESOURCE');
+    }
+
+    const passwordHash = await bcrypt.hash(requestDto.password, 10);
+
+    const insertedUser = await this.db.transaction(async (tx) => {
+      const [createUser] = await this.userRepository.createUser(
+        tx,
+        requestDto.email,
+        requestDto.nickName,
+        requestDto.gender,
+      );
+      await this.userRepository.createNativeUser(
+        tx,
+        createUser.id,
+        passwordHash,
+      );
+      return createUser;
+    });
+
+    if (insertedUser.authRole === 'LEFT') {
+      throw new CustomException('LOGIN_ERROR');
+    }
+
+    const { tokenResponse, refreshTokenExpiredAt } =
+      await this.authUtil.issueTokens(
+        insertedUser.id,
+        insertedUser.authRole,
+        false,
+      );
+
+    await this.userRepository.createRefreshToken(
+      this.db,
+      insertedUser.id,
+      tokenResponse.refreshToken,
+      refreshTokenExpiredAt,
+    );
+
+    const response = {
+      ...tokenResponse,
+      nickName: insertedUser.nickName,
+      gender: insertedUser.gender,
+      authRole: insertedUser.authRole,
+    };
+
+    this.log.log(`Sign Up successfully for user ID: ${insertedUser.id}`);
+
+    return plainToInstance(SignInResponseDto, response, {
+      excludeExtraneousValues: true,
+    });
   }
 }
